nixos: Use systemd.mounts instead of fileSystems for bind mounts
This removes the need for the `home` option, as the home directory can
be deduced automatically. This was previously not possible due to a
long-standing issue in nixpkgs:
https://github.com/NixOS/nixpkgs/issues/24570.

It also greatly simplifies the neededForBoot directory creation, since
it's either handled automatically by systemd in initrd or easily done
in the postMountCommands.

Fixes #179