home-manager: Replace module with an interface to the NixOS module
Since the NixOS module is much more feature complete and able to
leverage real bind mounts, make the Home Manager module another
interface to the NixOS module. The required functionality and more is
already implemented for the `users` submodule option.

This should close a whole bunch of issues related to the Home Manager
module and massively lessen the maintenance burden of the project.

The main caveat is that the NixOS module has to be imported into the
same configuration as Home Manager for this to work. I can't find any
good reason to use the Home Manager module exclusively, but if you
have one - do tell.